Few changes and refactoring
