+++ packages/opencensus-web-core/src/trace/model/tracing.ts
@@ -39,12 +39,6 @@ export class Tracing implements webTypes.Tracing {
     return this.singletonInstance || (this.singletonInstance = new this());
   }
 
-  constructor() {
-    // register the noop exporter to have an span event listener,
-    // helpful for testing.
-    this.tracer.registerSpanEventListener(this.exporter);
-  }
-
   /** Sets tracer and exporter config. */
   start(config?: webTypes.Config): webTypes.Tracing {
     this.tracer.start(config || {});
