+++ packages/opencensus-web-instrumentation-zone/src/on-page-interaction-stop-watch.ts
