+++ packages/opencensus-web-instrumentation-zone/test/index.ts
@@ -16,5 +16,5 @@
 
 // This file is an entry point for the webpack test configuration, so this
 // should import from all test files.
-import './test-on-page-interaction';
+import './test-on-page-interaction-stop-watch';
 import './test-interaction-tracker';
